### 2024-09-07 #7197 @editorialbot
[PRE REVIEW]: Katsu: Integrated polarimetry and polarization simulation
**Submitting author:** @Jashcraf (<a href="http://orcid.org/0000-0001-5082-7442">Jaren Ashcraft</a>)
**Repository:** https://github.com/Jashcraf/katsu
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
