### 2024-09-07 #7197 @dfm
In the meantime, there's some initial work to do on the repository:

1. Please add a license.
2. You should clean up all the temporary files. Things like `.DS_Store`, `.ipynb_checkpoints`, `*.egg-info`, `docs/_build`, and `__pycache__` shouldn't be checked into a repo. You can use a `.gitignore` to help with that.

Thanks!
