### 2024-09-08 6eccd15
Added License, in progress support for spatially-varying fits
