### 2024-09-08 428eee4
Deleting all temporary files
