+++ katsu.egg-info/requires.txt
@@ -1,3 +1,0 @@
-numpy
-scipy
-pyserial
