### 2024-09-14 deea56f
adding Smith polarimeter optimization tutorial, not in website yet
