### 2024-10-13 #7197 @dfm
:wave: @scottprahl, @benjaminpope, @arendMoerman — would any of you be willing to review this submission for The Journal of Open Source Software? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
