### 2024-10-19 #7375 @editorialbot
[REVIEW]: Katsu: Integrated polarimetry and polarization simulation
**Submitting author:** @Jashcraf (<a href="http://orcid.org/0000-0001-5082-7442">Jaren Ashcraft</a>)
**Repository:** https://github.com/Jashcraf/katsu
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.1
