### 2024-10-21 c01c9c1
trying different numpy version for tests
