### 2024-10-21 ae2e064
Update to fix a couple errors in the jax depolarizer decomposition, some in-progress demos
