+++ docs/requirements.txt
@@ -1,4 +1,4 @@
-numpy>=2.1.0
+numpy>=1.23.5
 tox>=3.24.3
 pytest>=6.2.5
 pytest-cov>=2.12.1
