### 2024-10-21 2daa6bc
updated power
