+++ paper.bib
@@ -44,7 +44,8 @@ abstract = {The dual-rotating-retarder configuration is one of the most common f
   volume={672},
   pages={A121},
   year={2023},
-  publisher={EDP Sciences}
+  publisher={EDP Sciences},
+  doi={	https://doi.org/10.1051/0004-6361/202245651}
 }
 
 @misc{gaudi2020habitable,
@@ -53,7 +54,8 @@ abstract = {The dual-rotating-retarder configuration is one of the most common f
       year={2020},
       eprint={2001.06683},
       archivePrefix={arXiv},
-      primaryClass={astro-ph.IM}
+      primaryClass={astro-ph.IM},
+      doi={https://doi.org/10.48550/arXiv.2001.06683}
 }
 
 @inproceedings{Will_polarization_luvoir,
@@ -198,6 +200,7 @@ abstract = {Astronomical instruments to detect exoplanets require extreme wavefr
       archivePrefix={arXiv},
       primaryClass={astro-ph.IM},
       url={https://arxiv.org/abs/2406.19616}, 
+      doi={https://doi.org/10.1117/12.3019204}
 }
 
 @misc{gromit,
@@ -215,7 +218,8 @@ title = {{Half-wave plate characterizations for the Keck NIRC2 polarimetry mode}
 organization = {International Society for Optics and Photonics},
 publisher = {SPIE},
 keywords = {NIRC2, polarimetric imaging, waveplate},
-year = {2023}
+year = {2023},
+doi = {https://doi.org/10.1117/12.3019556}
 }
 
 @software{jax2018github,
