### 2024-10-22 #7375 @Jashcraf
@dfm Fixed most of the DOIs, it looks like editorialbot doesn't like pypol's zenodo doi. Can I keep that one as-is or should I write it differently in the paper.bib file?
