### 2024-10-22 #7375 @dfm
> @dfm Fixed most of the DOIs, it looks like editorialbot doesn't like pypol's zenodo doi. Can I keep that one as-is or should I write it differently in the paper.bib file?

I think the DOI you want there is 10.5281/zenodo.8358112. The zenodo URL isn't a valid DOI, but the appropriate DOI is listed on that page. Is that the one you were talking about?
