### 2024-10-23 #7375 @benjaminpope
Student theses are due this week, so I might do this in bits, if that ok, as time becomes available.

Some quick thoughts on functionality:

### Functionality

- **Installation:** Does installation proceed as outlined in the documentation?
    - The documentation doesn't outline installation! This is just in the GitHub readme - you should add it into the docs landing page. Also - no PyPI? `katsu` is available on PyPI and I suggest you grab it fast!
    - needs to require: `[jax,dLux,prysm]`
- **Functionality:** Have the functional claims of the software been confirmed?
    - All notebooks in `docs` executed in clean `conda` environment.
    - In `CalibratingPolarimeters.ipynb`, bug at `from katsu.polarimetry import broadcasted_full_mueller_polarimetry`
    - In `FisherInformationDRRP.ipynb`, shape error at `plt.imshow(fisher_matrix, cmap='RdBu_r')`
    - in `DRRP_ADOptimization.ipynb`, at `print_results(results)`, I get a `TypeError: jacrev requires real- or complex-valued inputs (input dtype that is a sub-dtype of np.inexact), but got int64. If you want to use Boolean- or integer-valued inputs, use vjp or set allow_int to True.`
