### 2024-10-23 Jashcraf/katsu#30 @arendMoerman
Clearer link to documentation
This issue is part of my JOSS review: openjournals/joss-reviews#7375

Currently, the only way documentation for `katsu` can be accessed (as far as I can see) is by clicking the
![image](https://github.com/user-attachments/assets/c63bf8a4-65a0-4f60-b300-4f37d1f0d2c0)
button, which is not the most obvious/normative way to do this.

It would be nice if the documentation link is clearly and explicitly stated in the README. Also, in the "About" section of the  repository, it would be nice to place a link to the documentation. As an example, the [scipy](https://github.com/scipy/scipy) repository implements both a clear link to the documentation in the README, and displays the link to the Scipy website in the "About" section of the repo.
