### 2024-10-29 e294475
Clearer link to documentation in the README, added logo, fixed zenodo doi
