+++ paper.bib
@@ -182,7 +182,7 @@ abstract = {Astronomical instruments to detect exoplanets require extreme wavefr
   howpublished = {\url{https://github.com/scottprahl/pypolar}},
   title        = {Pypolar},
   year         = {2024},
-  doi          = {https://zenodo.org/badge/latestdoi/128805296}
+  doi          = {10.5281/zenodo.8358112}
 }
 
 @misc{key,
