### 2024-10-29 Jashcraf/katsu#30 @Jashcraf
Made the requested changes by highlighting docs link as a subsection in the README, and adding the link to the "About" section of the repo. Thanks for the reccomendation!

Change made in [e294475](https://github.com/Jashcraf/katsu/commit/e294475e7c70c3f8460712946965c2c374eb829e)
