### 2024-10-29 #7375 @Jashcraf
@dfm thanks for the correction! Updated that citation in [e294475](https://github.com/Jashcraf/katsu/commit/e294475e7c70c3f8460712946965c2c374eb829e)

Addressed @arendMoerman 's comment in the issue they made on the Katsu repo

@benjaminpope I'll jump on grabbing Katsu on PyPI! Thanks for pointing out those specific notebooks - those are "in progress" docs, so they aren't included in the [index.rst](https://github.com/Jashcraf/katsu/blob/main/docs/index.rst) that shows up on [the readthedocs website](katsu.readthedocs.io). Would it be better to have them in a separate directory, titled something like "in progress" or "experimental"?
