### 2024-10-29 Jashcraf/katsu#30 @arendMoerman
Thank you for implementing the changes. Locating the docs is a lot clearer now!
