### 2024-10-29 #7375 @Jashcraf
@benjaminpope just went and verified that all of the notebooks included in the index.rst work
