### 2024-10-30 Jashcraf/katsu#31 @arendMoerman
JOSS paper summary and statement of need confusion
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7375

The JOSS reviewer guidelines specify quite clearly what the paper sections should contain:

<b>Summary</b>: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

<b>A statement of need</b>: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

Currently, using above definitions, the paper reads as if the summary contains (a section of) the statement of need. Conversely, the statement of need section seems to contain some of the high-level overview which seems more appropriate in the summary section, by the JOSS reviewers guideline.

Could you please reformulate the summary and statement of need sections in the JOSS paper so that they align more closely with the JOSS reviewer guideline definitions?
