### 2024-10-30 Jashcraf/katsu#32 @arendMoerman
JOSS paper ambiguous use of mathematical symbol for Mueller matrix and depolarizer
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7375

In the paper, there is one mathematical equation describing the formalism of the Mueller matrix decomposition, where the full matrix $$M$$ is decomposed into a depolarizer, diattenuator, and retarder:
$$M=M M_\mathrm{R} M_\mathrm{D}.$$
As it is now, the symbol $$M$$ is used for both the total matrix and the depolarizer constituent of the decomposition, which is ambiguous.

I believe this to be just a typo in the paper however, as in the [online documentation](https://katsu.readthedocs.io/en/latest/notebooks/PolarDecomposition.html) the symbol $$M_\Delta$$ is used for the depolarizer, which would break the ambiguity and resolve this issue.

Could you please fix this in the JOSS paper please?
