### 2024-10-30 Jashcraf/katsu#32 @Jashcraf
Thank you for the catch! I think it may be a compilation issue? The $M_{\Delta}$ is present in `paper.md`

https://github.com/Jashcraf/katsu/blob/e294475e7c70c3f8460712946965c2c374eb829e/paper.md?plain=1#L57

And shows up in the markdown preview
![image](https://github.com/user-attachments/assets/25970ec6-a738-4dc7-ae3b-5e228733da3f)

Looking at the Github action I found the source of the error
```
[INFO] [makePDF] LaTeX run number 1
Latexmk: ====Undefined refs and citations with line #s in .tex file:
Missing character: There is no 𝛥 (U+1D6E5) in font [NewCM10-Bold.otf]:mode=ba
Missing character: There is no 𝛥 (U+1D6E5) in font [NewCM10-Bold.otf]:mode=base;script=latn;language=dflt;!
Missing character: There is no 𝛥 (U+1D6E5) in font [NewCM10-Bold.otf]:mode=ba
Missing character: There is no 𝛥 (U+1D6E5) in font [NewCM10-Bold.otf]:mode=base;script=latn;language=dflt;!
Missing character: There is no 𝛥 (U+1D6E5) in font [NewCM10-Bold.otf]:mode=ba
Missing character: There is no 𝛥 (U+1D6E5) in font [NewCM10-Bold.otf]:mode=base;script=latn;language=dflt;!
Missing character: There is no 𝛥 (U+1D6E5) in font [NewCM10-Bold.otf]:mode=ba
```

It may not be present because the Delta is inside the `\mathbf{*}`, I'll move the Delta outside to see  if that fixes it.
