### 2024-10-30 9fee486
fixing typo in depolarizer equation
