+++ paper.md
@@ -52,9 +52,9 @@ Polarimetric characterization in the laboratory is critical to the performance o
 # Statement of need
 `Katsu` is an open-source Python package to address the need for polarimetric characterization of astronomical systems for the next generation of astronomy. It contains simple routines for the simulation of Mueller matrices and Stokes vectors. One area where `Katsu` is distinct from other excellent packages capable of Mueller calculus simulation (e.g., Pypol [@pypol]) is its emphasis on broadcasted matrix calculations. All Mueller matrices available in `katsu.mueller` take a `shape` keword that append an arbitrary number of axes to the front of the initialized array, with the final two axes containing the Mueller matrix. This functionality is critical for accelerated computating on spatial data, which enables the direct measurement of polarization aberrations in the lab [@ashcraft2024]. `Katsu` also features a `polarimetry` module containing the data reduction routines for single-rotating retarder (SRR) Stokes polarimeters and dual-rotating retarder (DRR) Mueller polarimeters. In the pursuit of open-source instrumentation in the laboratory, `Katsu` also supports an interface to the Newport Agilis series rotation mounts in the `motion` module to assist with performing polarimetry with rotating retarders. 
 
-To interpret the measurements made in the lab, `Katsu` features the polar decomposition of Mueller matrices published by Lu and Chipman [@Lu:96]. This decomposes a Mueller matrix $\mathbf{M}$ into its constituent depolarizer $\mathbf{M_{\Delta}}$, diattenuator $\mathbf{M_{D}}$, and retarder $\mathbf{M_{R}}$, as shown in the following Equation,
+To interpret the measurements made in the lab, `Katsu` features the polar decomposition of Mueller matrices published by Lu and Chipman [@Lu:96]. This decomposes a Mueller matrix $\mathbf{M}$ into its constituent depolarizer $\mathbf{M}_{\Delta}$, diattenuator $\mathbf{M_{D}}$, and retarder $\mathbf{M_{R}}$, as shown in the following Equation,
 
-$$\mathbf{M} = \mathbf{M_{\Delta}}\mathbf{M_{R}}\mathbf{M_{D}}. $$
+$$\mathbf{M} = \mathbf{M}_{\Delta}\mathbf{M_{R}}\mathbf{M_{D}}. $$
 
 This function is critical for separating depolarization from the Mueller matrix, which enables the integration of polarization aberration in the laboratory into simulated data (generated by a polarization ray tracer, e.g. `Poke` [@Ashcraft_2023], `Pyastropol` [@Pruthvi.2020]). `Katsu` also adopts the flexible interchangeable backend system of `prysm` [@Dube2019;@Dube:22] for hot-swappable `numpy`-like backends [@harris2020array] including `cupy` for GPU-accelerated computing [@cupy_learningsys2017] and `jax` for automatic differentiation [@jax2018github].
 
