### 2024-10-30 Jashcraf/katsu#32 @Jashcraf
![image](https://github.com/user-attachments/assets/d3834bc6-7eac-4360-b98d-990b0754926c)

Looks like it worked!
