### 2024-10-30 2067831
PyPI packaging error, fixing invalid URL
