### 2024-10-30 46f113a
Added pip to install instructions
