### 2024-10-30 #7375 @Jashcraf
@benjaminpope Katsu added to PyPI!
https://pypi.org/project/katsu/
