### 2024-11-01 bedeee2
reformulation of sections to better fit JOSS review guidelines
