+++ paper.md
@@ -11,7 +11,7 @@ authors:
   - name: Jaren N. Ashcraft
     orcid: 0000-0001-5082-7442
     corresponding: true
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+    affiliation: "1, 2, 3" # (Multiple affiliations must be quoted)
   - name: Ewan S. Douglas
     orcid: 0000-0002-0813-4308
     affiliation: 2
@@ -47,10 +47,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Polarimetric characterization in the laboratory is critical to the performance of astronomical instrumentation. The next generation of astronomical observatories has identified that wavefront errors induced by polarization, called _polarization aberrations_, could be a limiting factor in direct exoplanet imaging. Ample modeling has been done at the observatory level to understand the nominal polarization aberrations that arise in the telescope [@anche2023polarization;@gaudi2020habitable;@Will_polarization_luvoir] but less work has been done to characterize instrumentation in the laboratory.
-
-# Statement of need
-`Katsu` is an open-source Python package to address the need for polarimetric characterization of astronomical systems for the next generation of astronomy. It contains simple routines for the simulation of Mueller matrices and Stokes vectors. One area where `Katsu` is distinct from other excellent packages capable of Mueller calculus simulation (e.g., Pypol [@pypol]) is its emphasis on broadcasted matrix calculations. All Mueller matrices available in `katsu.mueller` take a `shape` keword that append an arbitrary number of axes to the front of the initialized array, with the final two axes containing the Mueller matrix. This functionality is critical for accelerated computating on spatial data, which enables the direct measurement of polarization aberrations in the lab [@ashcraft2024]. `Katsu` also features a `polarimetry` module containing the data reduction routines for single-rotating retarder (SRR) Stokes polarimeters and dual-rotating retarder (DRR) Mueller polarimeters. In the pursuit of open-source instrumentation in the laboratory, `Katsu` also supports an interface to the Newport Agilis series rotation mounts in the `motion` module to assist with performing polarimetry with rotating retarders. 
+`Katsu` is an open-source Python package to address the need for polarimetric characterization of astronomical systems for the next generation of astronomy. It contains simple routines for the simulation of Mueller matrices and Stokes vectors. One area where `Katsu` is distinct from other excellent packages capable of Mueller calculus simulation (e.g., Pypol [@pypol]) is its emphasis on broadcasted matrix calculations. All Mueller matrices available in `katsu.mueller` take a `shape` keword that append an arbitrary number of axes to the front of the initialized array, with the final two axes containing the Mueller matrix. This functionality is critical for accelerated computating on spatial data, which enables the direct measurement of polarization aberrations in the lab [@ashcraft2024]. `Katsu` also features a `polarimetry` module containing the data reduction routines for single-rotating retarder (SRR) Stokes polarimeters and dual-rotating retarder (DRR) Mueller polarimeters. In the pursuit of open-source instrumentation in the laboratory, `Katsu` supports an interface to the Newport Agilis series rotation mounts in the `motion` module to assist with performing polarimetry with rotating retarders. 
 
 To interpret the measurements made in the lab, `Katsu` features the polar decomposition of Mueller matrices published by Lu and Chipman [@Lu:96]. This decomposes a Mueller matrix $\mathbf{M}$ into its constituent depolarizer $\mathbf{M}_{\Delta}$, diattenuator $\mathbf{M_{D}}$, and retarder $\mathbf{M_{R}}$, as shown in the following Equation,
 
@@ -58,9 +55,12 @@ $$\mathbf{M} = \mathbf{M}_{\Delta}\mathbf{M_{R}}\mathbf{M_{D}}. $$
 
[…]
