### 2024-11-01 Jashcraf/katsu#31 @Jashcraf
Reorganized sections of the paper to better fit the JOSS reviewer guideline definitions in [bedeee2](https://github.com/Jashcraf/katsu/commit/bedeee2ac1d4abf9580a3deae0f9be53e39436b1)
