### 2024-11-03 #7375 @benjaminpope
Great - good it's on PyPI.

Anyway - moving on with the review.

 **A statement of need**:
- This section of the paper needs a bit of work. It is maybe not 100% clear from reading it what `katsu` does and how it relates to other optics packages. I would suggest you have another half paragraph right at the top situating it in the context of other optics packages (eg you use dLux and prysm - cite and briefly say what these and any other relevant packages do). Then also explain, in one or two sentences, what Mueller matrices and Stokes vectors are - even just words to the effect that Stokes vectors are the appropriate representation for arbitrarily polarized light, and Mueller matrices are the objects that transform Stokes vectors when they propagate and interact with optical surfaces and media. It should be made clear, very briefly, that this is non-trivial. The existing text is too high level, and caters only to experts. The intro to Mueller calculus section of the docs is good but still too high level - at the level of a couple of paragraphs and wiki links, nothing too onerous, I'd suggest that the docs explain what this package is for at a very basic level.

**Installation instructions**:
- Can you add an Installation section to the readthedocs.io page?
- In fact, it would be nice to have a landing page on the docs with a bit more explanation of what this is for.

**Example usage**:
- The examples on the code's pure-simulation use good, though the intro to Mueller one could be expanded to be even more basic about Stokes if you have a chance.
- The example on motion control depends too much on external examples: is it possible to expand on this?

**Functionality documentation**:
- some of the functions still need docstrings - eg all the `katsu.motion` objects have incomplete docstrings on their methods and init. Some of the dostring annotation of parameters and types eg in `katsu.polarimetry.q_ultimate_polarimetry` has failed to render correctly in the readthedocs.io site - read through the whole of https://katsu.readthedocs.io/en/latest/katsu.html and look for which ones have this bug.

**State of the Field**:
- The paper should probably explain a little more what PyPol is, and also situate this in the optics literature - the package makes use of both dLux and prysm in example notebooks, perhaps mention and describe this connection and these use cases?
