### 2024-11-04 c3f6a4a
Adding installation instructions to landing page
