### 2024-11-05 90a96bf
moving install instructions
