### 2024-11-12 Jashcraf/katsu#31 @arendMoerman
Thank you for re-arranging the sections, the structure is indeed more reflective of the standards now.
