### 2024-11-12 Jashcraf/katsu#33 @arendMoerman
Better structure/layout and cleanup of documentation
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7375

Looking at the documentation, the contents are presented as a linear list, where the subject increments from purely theoretical and understanding-based ([Intro to Mueller Calculus](https://katsu.readthedocs.io/en/latest/notebooks/MuellerCalculus.html#Intro-to-Mueller-Calculus)) to quite practical and applied ([Dual Rotating Retarder Mueller Polarimetry](https://katsu.readthedocs.io/en/latest/notebooks/FullMuellerExample.html)), back to quite theoretical understanding-based ([Interpretation of Mueller Matrices based on the Polar decomposition, or the Lu-Chipman decomposition](https://katsu.readthedocs.io/en/latest/notebooks/PolarDecomposition.html#Interpretation-of-Mueller-Matrices-based-on-the-Polar-decomposition,-or-the-Lu-Chipman-decomposition)).

It would be better if you could structure the documentation in a way that separates these different subjects, such as described in [the documentation system](https://docs.divio.com/documentation-system/). This will make `katsu` easier to grasp for new users by providing a clear set of tutorials, whilst also (for example) provide more advanced users with a clear set of theoretical background information. In short, by dividing the documentation into these 'chunks', `katsu` will be easier to traverse for different levels of users.

Also, in general, the documentation should be cleaned up. For example, the [Single Rotating Retarder Stokes Polarimetry](https://katsu.readthedocs.io/en/latest/notebooks/FullStokesExample.html) contains a Python import stub at the top of the page, before the title. These little slip-ups should be fixed as well. Also keep an eye out for typos and spelling mistakes (e.g. 'computpe' in [this page](https://katsu.readthedocs.io/en/latest/notebooks/FullMuellerExample.html))
