### 2024-11-13 Jashcraf/katsu#34 @arendMoerman
No instructions on how to run automated tests in documentation
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7375

As is described in the reviewer guidelines, the documentation should contain the following:

<b>Automated tests</b>: Are there automated tests or manual steps described so that the functionality of the software can be verified?

To the best of my knowledge, there are no instructions or steps that describe how the tests can be run.
I do see that tests have been written and are present in the repository, but the documentation still needs to explicitly state how someone would run these tests.
