### 2024-11-13 Jashcraf/katsu#35 @arendMoerman
No community guidelines in documentation
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7375

The reviewer guidelines specify that the documentation should contain community/contributor guidelines:

<b>Community guidelines</b>: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Currently, and to the best of my knwoledge, there is a small paragraph at the end of the [Building the polarimetric data reduction from Q Tutorial](https://katsu.readthedocs.io/en/latest/notebooks/Q_DataReduction.html) where some mention is made of contributing.

The community/contributor guidelines need to have a separate page in the docs, so that new users can easliy find them. Also, the guidelines need to be more extensive than what is written on the above-mentioned page. On the README of the repo I see that some guidelines have been written. These would serve as a good starting point for the guidelines page in the actual documentation.
