### 2024-12-04 fd63e5e
Making some changes to demos, added max likelihood experiment
