### 2024-12-11 0e30505
trying adding the contents flag
