### 2024-12-11 Jashcraf/katsu#33 @Jashcraf
Just went through and did a restructure, my interpretation of the topics in the Documentation system are:
- Tutorials: more theoretical and understanding-based
- How-tos: Solving specific problems
- Explanations: practical and applied problems / deep-dives

Here's how they look on the sidebar
![image](https://github.com/user-attachments/assets/f5fa3f1b-c4cf-454e-a8ab-992a7f5b0a4f)

I'll go through and do some cleanup of the typos and python snippets up top now
