### 2024-12-19 87f6d13
some typo correction, addition of community guidelines and guidance for automated testing
