### 2024-12-19 Jashcraf/katsu#34 @Jashcraf
Added this to the installation section of the documentation page, since it is a good way to verify installation
