### 2024-12-19 Jashcraf/katsu#35 @Jashcraf
I added a Contribution section, including the contributing section from the README and a Code of Conduct adapted from The Contributor Covenant. Does this address community/contributor guidelines?
