### 2024-12-20 Jashcraf/katsu#35 @arendMoerman
Thank you for adding the community guidelines to the documentation. This makes it a lot clearer for people who wish to contribute! I think this is now taken care of so I will close this issue now.
