### 2024-12-20 Jashcraf/katsu#34 @arendMoerman
Thank you for adding the paragraph on automated testing. It seems that `pytest` is required for running them. Could you also specify that in the docs? It is not mentioned in the requirements so this should be added, at least as some optional requirement for running the tests.

After installing `pytest`, I ran the mentioned command, and it ran. However, a lot of tests in `test_mueller.py` and one in `test_katsu_math.py` were skipped. Upon inspecting the tests, it seems to skip if `jax` is not installed. I checked and `jax` is not mentioned in the requirements. After installing it, no tests were skipped and everything was fine, except for the following warnings:
```
tests/test_mueller.py::test_decompose_depolarizer_jax
  /home/arend/Venvs/katsu/lib/python3.11/site-packages/jax/_src/ops/scatter.py:92: FutureWarning: scatter inputs have incompatible types: cannot safely cast value from dtype=complex128 to dtype=float64 with jax_numpy_dtype_promotion='standard'. In future JAX releases this will result in an error.
    warnings.warn(

tests/test_mueller.py::test_decompose_depolarizer_jax
  /home/arend/Venvs/katsu/lib/python3.11/site-packages/jax/_src/ops/scatter.py:136: ComplexWarning: Casting complex values to real discards the imaginary part
    return lax_internal._convert_element_type(out, dtype, weak_type)
```
Does not seem too serious (some casting warnings), except for that the first warning will result in an error in future `jax` versions. Maybe worth a look.

I think this all ties back into the previous point: there should also be a list of optional requirements for running the tests, or for enabling `jax`. Adding this would make the docs even more complete.
