### 2024-12-20 Jashcraf/katsu#36 @arendMoerman
Documentation lacks API reference
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7375

It seems that the documentation still lacks a comprehensive API reference, or at least a clear page for this.
Upon inspection, it seems that the landing page contains an 'Indices and tables' section at the bottom of the page. Clicking on the 'index' bullet point reveals an alphabetically ordered list that resembles an API reference. However, it is easy to miss this page because it is not listed in the sidemenu and hence has to be listed more explicitly.

Also, it would be nice if the API reference could be ordered in a more sensible way. For example, separate subpages could be made for the `katsu_math` , `motion`, `mueller`, etc. modules. Then, alphabetical ordering could be applied within such a single subpage. This would enhance the readability and ease of use of the API reference.

As an example, have a look at the [SciPy](https://docs.scipy.org/doc/scipy/reference/index.html#scipy-api) API reference. Here they quite nicely implemented the aforementioned subdivision in the API reference.
