### 2024-12-20 Jashcraf/katsu#33 @arendMoerman
Thanks for implementing this. It looks a lot more structured now.
It still seems that some artefacts persist (e.g., the import stub on [this page of the docs](https://katsu.readthedocs.io/en/latest/notebooks/explanation/DualChannelPolarimetryIntro.html#Single-Differencing)). I will assume these will be taken care of though, so I will close this issue now.
