### 2024-12-20 Jashcraf/katsu#36 @Jashcraf
Thanks for the direction to the SciPy API reference!
