### 2025-01-15 1c4f06f
Adding optional dependencies to installation page
