### 2025-01-15 Jashcraf/katsu#34 @Jashcraf
Thanks for pointing these errors out! I included the following section in the installation page
