### 2025-01-15 Jashcraf/katsu#34 @arendMoerman
Great! These issues seem addressed now so I will close this.
