### 2025-01-15 b477cc9
try change api ref toctree depth
