+++ docs/index.rst
@@ -21,6 +21,7 @@ Summary
 .. toctree::
    :maxdepth: 2
    :caption: API Reference
+   
    modules
 
 .. toctree::
