+++ docs/modules.rst
@@ -1,7 +1,8 @@
 katsu API Reference
 =====
 
-**Importing from Katsu**
+Importing from Katsu
+--------------------
 Imports in katsu are done by calling the appropriate submodules. The code snippet below shows how to import the method to create a linear diattenuator
 
 .. code-block:: python
@@ -10,7 +11,8 @@ Imports in katsu are done by calling the appropriate submodules. The code snippe
 
 Methods or classes with an underscore "_" preceding the name indicates that this is not a public method or class, and therefore its direct use is not reccomended. For a complete listing of the methods available in Katsu, see the following API reference:
 
-**API Reference**
+API Reference
+-------------
 
 .. toctree::
    :maxdepth: 4
