+++ docs/modules.rst
@@ -1,5 +1,6 @@
 katsu API Reference
-=====
+===================
+
 
 Importing from Katsu
 --------------------
@@ -10,6 +11,7 @@ Imports in katsu are done by calling the appropriate submodules. The code snippe
    from katsu.mueller import linear_diattenuator
 
 Methods or classes with an underscore "_" preceding the name indicates that this is not a public method or class, and therefore its direct use is not reccomended. For a complete listing of the methods available in Katsu, see the following API reference:
+
 
 API Reference
 -------------
