### 2025-01-15 Jashcraf/katsu#36 @Jashcraf
Added a page to the API reference, which you can see here
https://katsu.readthedocs.io/en/latest/modules.html#
