+++ docs/motion.rst
@@ -1,0 +1,7 @@
+katsu.motion module
+-------------------
+
+.. automodule:: katsu.motion
+   :members:
+   :undoc-members:
+   :show-inheritance:
