+++ docs/mueller.rst
@@ -1,0 +1,7 @@
+katsu.mueller module
+--------------------
+
+.. automodule:: katsu.mueller
+   :members:
+   :undoc-members:
+   :show-inheritance:
