+++ docs/polarimetry.rst
@@ -1,0 +1,7 @@
+katsu.polarimetry module
+------------------------
+
+.. automodule:: katsu.polarimetry
+   :members:
+   :undoc-members:
+   :show-inheritance:
