+++ docs/katsu_math.rst
@@ -1,0 +1,7 @@
+katsu.katsu_math module
+-----------------------
+
+.. automodule:: katsu.katsu_math
+   :members:
+   :undoc-members:
+   :show-inheritance:
