### 2025-01-15 Jashcraf/katsu#36 @Jashcraf
Alright each module in the API reference has its own page now, ex: katsu_math

https://katsu.readthedocs.io/en/latest/katsu_math.html
