### 2025-01-16 Jashcraf/katsu#36 @arendMoerman
Thanks for improving the API reference! I consider this resolved, so will close the issue now.
