### 2025-01-20 d8ff60d
Fixing docstrings, some cleanup surrounding functions in .mueller and .polarimetry
