+++ katsu/mueller.py
@@ -276,10 +276,6 @@ def linear_retarder(a, r, shape=None):
         M = M.at[..., 3, 3].set(np.cos(r))
     
     else:
-
-        # if M.ndim > 4:
-        #     a = a[..., np.newaxis]
-        #     r = r[..., np.newaxis]
             
         # First row
         M[..., 0, 0] = 1.
@@ -304,7 +300,7 @@ def linear_retarder(a, r, shape=None):
 def linear_diattenuator(a, Tmin, Tmax=1, shape=None):
     """returns a homogenous linear diattenuator
 
-    CLY 6.54
+    See Equation 6.54 in CLY
 
     Parameters
     ----------
@@ -369,13 +365,6 @@ def linear_diattenuator(a, Tmin, Tmax=1, shape=None):
 
     else:
 
-        # if M.ndim > 4:
-        #     A = A[..., np.newaxis]
-        #     B = B[..., np.newaxis]
-        #     C = C[..., np.newaxis]
-        #     cos2a = cos2a[..., np.newaxis]
-        #     sin2a = sin2a[..., np.newaxis]
-
         # first row
         M[..., 0, 0] = A
         M[..., 0, 1] = B*cos2a
@@ -399,42 +388,35 @@ def linear_diattenuator(a, Tmin, Tmax=1, shape=None):
 
     return M
 
-def wollaston(beam = 0, eta = 1):
-    """returns the 
+def wollaston(beam = 0, rotation=0., shape=None):
+    """Method to construct the Mueller matrix of a Wollaston, 
+    Functionally just a hand-hold wrapper for linear_polarizer
 
     Parameters
     ----------
-    angle : float, or numpy.ndarray
-        angle of the transmission axis w.r.t. horizontal in radians. If numpy
-        array, must be the same shape as `shape`
-    a : float or numpy.ndarray
-        depolarization of Q
-    a : float or numpy.ndarray
-        depolarization of U
-    a : float or numpy.ndarray
-        depolarization of V
+    beam : int, or str
+        'Channel' of wollaston prism, by default 0, which returns the
+        Mueller matrix for the horizontally-polarized beam
+    rotation : float, optional
+        Rotation of the Wollaston prism with respect to the laboratory's
+        horizontal axis, by default 0
     shape : list, optional
         shape to prepend to the mueller matrix array, see `_empty_mueller`.
-        By default Nonex
+        By default None
 
     Returns
     -------
-    _type_
-        _description_
+    numpy.ndarray
+        Mueller matrix of the chosen linear polarizer channel
     """
-    # If the ordinary beam then sign = 1
-    if beam == 0:
-        sign = eta
-    # If the extraordinary beam then sign = -1
+
+    # Ordinary beam
+    if (beam == 0) or (beam=="ordinary"):
+        return linear_polarizer(rotation, shape=shape)
+    
+    # Extraordinary beam
     else:
-        sign = -eta
-
-    M = 0.5 * np.array([[1, sign, 0, 0],
-        [sign, 1, 0, 0],
-        [0, 0, 0, 0],
-        [0, 0, 0, 0]])
-
-    return M
+        return linear_polarizer(rotation + np.pi/2, shape=shape)
 
 def depolarizer(angle, a, b, c, shape=None):
     """returns a diagonal depolarizer
@@ -446,9 +428,9 @@ def depolarizer(angle, a, b, c, shape=None):
[…]
