+++ katsu/polarimetry.py
@@ -1,6 +1,12 @@
 from .mueller import linear_retarder, linear_polarizer, linear_diattenuator, _empty_mueller, decompose_retarder, wollaston
 from .katsu_math import broadcast_kron, broadcast_outer, condition_number, RMS_calculator, propagated_error, np
-from scipy.optimize import curve_fit
+from scipy.optimize import curve_fit, minimize
+
+# Define the identity matrix and other matrices which are useful for the Mueller calculus
+M_identity = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, 1]])
+A = np.array([1, 0, 0, 0])
+B = np.array([[1], [0], [0], [0]])
+C = np.array([0, 1, 0, 0])
 
 
 def drrp_data_reduction_matrix(Mg, Ma, invert=False):
@@ -33,8 +39,6 @@ def drrp_data_reduction_matrix(Mg, Ma, invert=False):
     
     else:
         return Wmat
-
-
 
 
 def full_mueller_polarimetry(thetas, power, angular_increment,
@@ -133,6 +137,7 @@ def full_mueller_polarimetry(thetas, power, angular_increment,
 
     return M_meas.reshape([*M_meas.shape[:-1], 4, 4])
 
+
 def stokes_sinusoid(theta, a0, b2, a4, b4):
     """sinusoidal response of a single rotating retarder full stokes
     polarimeter.
@@ -156,6 +161,7 @@ def stokes_sinusoid(theta, a0, b2, a4, b4):
         sinusoidal response of the SRRP
     """
     return a0 + b2*np.sin(2*theta) + a4*np.cos(4*theta) + b4*np.sin(4*theta)
+
 
 def full_stokes_polarimetry(thetas, Sin=None, power=None, return_coeffs=False):
     """conduct a full stokes polarimeter measurement
@@ -220,6 +226,7 @@ def full_stokes_polarimetry(thetas, Sin=None, power=None, return_coeffs=False):
     else:
         return np.array([S0, S1, S2, S3])
 
+
 # TODO: Figure out a way to get the dual_channel_polarimetry_function to use
 # this generalized function instead of separate sinusoidal fitting functions
 def dual_channel_sinusoid(theta, I, Q, U, theta_2 = None, 
@@ -275,6 +282,7 @@ def dual_channel_sinusoid(theta, I, Q, U, theta_2 = None,
             return (Q * (np.cos(4 * theta) - np.cos(4 * theta_2)) + \
                 U * (np.sin(4 * theta) - np.sin(4 * theta_2)))
         
+
 def unnormalized_single_diff_sinusoid(theta, Q, U):
     """
     Calculate the unnormalized sinusoidal response for a single difference in 
@@ -297,6 +305,7 @@ def unnormalized_single_diff_sinusoid(theta, Q, U):
     """
     output_power = (Q * np.cos(4 * theta) + U * np.sin(4 * theta))
     return output_power
+
 
 def unnormalized_double_diff_sinusoid(theta_1, theta_2, Q, U):
     """
@@ -326,6 +335,7 @@ def unnormalized_double_diff_sinusoid(theta_1, theta_2, Q, U):
         U * (np.sin(4 * theta_2) - np.sin(4 * theta_1)))
     return output_power
 
+
 # TODO: Test the implentation of normalized single difference
 def normalized_single_diff_sinusoid(theta, I, Q, U):
     """
@@ -350,6 +360,7 @@ def normalized_single_diff_sinusoid(theta, I, Q, U):
     
     output_power = (Q * np.cos(4 * theta) + U * np.sin(4 * theta) / I)
     return output_power
[…]
