+++ katsu/katsu_math.py
@@ -122,10 +122,12 @@ M_identity = np.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, 1]])
 
 def RMS_calculator(calibration_matrix):
     """Calculates the root mean square (RMS) error of a matrix by comparing it to the identity matrix.
+
     Parameters
     ----------
     calibration_matrix : array
         4x4 matrix.
+
     Returns
     -------
     RMS : float
@@ -144,12 +146,17 @@ def RMS_calculator(calibration_matrix):
 def propagated_error(M_R, RMS):
     """Propogates error in the Mueller matrix to error in the extracted value of retardance. 
     Assumes the RMS error is the same for all elements of the matrix.
+
+    Parameters
+    ----------
     M_R : array
         4x4 Mueller matrix for a linear retarder
     RMS : float
         root mean square error of the Mueller matrix.
+
     Returns
     ------- 
-        float, error in the extracted retardance value in radians."""
+        float, error in the extracted retardance value in radians.
+    """
     x = np.trace(M_R)
     return 2*RMS/np.sqrt(4*x-x**2) # Value in radians
