### 2025-01-24 Jashcraf/katsu#37 @arendMoerman
Incomplete tutorials show up in docs
This issue is part of my JOSS review: https://github.com/openjournals/joss-reviews/issues/7375

It seems that the repository contains a set of documentation notebooks that are incomplete.
In principle it is fine to have documentation pages that are 'work-in-progress', but they do show up in the actual docs which is quite confusing.

I found this when I was checking the usage of `dlux` in the `katsu` package (as per the comment of the other reviewer) by looking it up in the searchbar and stumbled upon the [Dual Rotating Retarder Polarimetry with Fisher Information](https://katsu.readthedocs.io/en/latest/notebooks/incomplete/FisherInformationDRRP.html#Dual-Rotating-Retarder-Polarimetry-with-Fisher-Information) page, which I could not locate in the navigation sidebar. The same happened when I wanted to check where `prysm` was used, where I ended up on the page for [Calibration and Optimization with Spatially-varying Data](https://katsu.readthedocs.io/en/latest/notebooks/incomplete/SpatiallyVaryingData.html).

Upon scrolling through these pages, I got the feeling that they were less developed than the pages that are accessible from the navigation sidebar, which led me to find out that they were 'incomplete'.

In order to avoid confusion, could you prevent the 'incomplete' tutorials from showing up in the docs?
