### 2025-02-03 #7375 @arendMoerman
I still have an open issue regarding the documentation on the software repository. After this is addressed, and after the points raised by @benjaminpope regarding the software paper are addressed, I feel that I can wrap up my portion of the review.
