### 2025-02-14 810dc41
adding orphan directive to all in progress demos, removing a demo that is no longer relevant
