### 2025-02-14 Jashcraf/katsu#37 @Jashcraf
Moved to a development branch. When I search for the Fisher Information docs it shows up in my search history

But you can't actually navigate to the page

I'm assuming this has to do with readthedocs saving my search history, and that they won't appear anymore.
