### 2025-02-14 Jashcraf/katsu#37 @arendMoerman
Thank you for taking care of this. Now when I look for `dlux` or `prysm`, it shows no results so it looks like it worked just fine! Will close this issue now.
