### 2025-02-15 10556dc
paper updates in response to JOSS comments, first draft
