+++ paper.bib
@@ -48,6 +48,33 @@ abstract = {The dual-rotating-retarder configuration is one of the most common f
   doi={	https://doi.org/10.1051/0004-6361/202245651}
 }
 
+@software{poppy,
+       author = {{Perrin}, Marshall and {Long}, Joseph and {Douglas}, Ewan and {Sivaramakrishnan}, Anand and {Slocum}, Christine and {others}},
+        title = "{POPPY: Physical Optics Propagation in PYthon}",
+ howpublished = {Astrophysics Source Code Library, record ascl:1602.018},
+         year = 2016,
+        month = feb,
+          eid = {ascl:1602.018},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2016ascl.soft02018P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{hcipy,
+       author = {{Por}, Emiel H. and {Haffert}, Sebastiaan Y. and {Radhakrishnan}, Vikram M. and {Doelman}, David S. and {van Kooten}, Maaike and {Bos}, Steven P.},
+        title = "{High Contrast Imaging for Python (HCIPy): an open-source adaptive optics and coronagraph simulator}",
+    booktitle = {Adaptive Optics Systems VI},
+         year = 2018,
+       editor = {{Close}, Laird M. and {Schreiber}, Laura and {Schmidt}, Dirk},
+       series = {Society of Photo-Optical Instrumentation Engineers (SPIE) Conference Series},
+       volume = {10703},
+        month = jul,
+          eid = {1070342},
+        pages = {1070342},
+          doi = {10.1117/12.2314407},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018SPIE10703E..42P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @misc{gaudi2020habitable,
       title={The Habitable Exoplanet Observatory (HabEx) Mission Concept Study Final Report}, 
[…]
