### 2025-02-15 5e4ab64
fixing errors in citation render
