### 2025-02-15 #7375 @Jashcraf
Alright I believe I'm caught up! Have some responses inline below:
- addressed @arendMoerman 's comment on the incomplete docs showing as searchable in katsu.readthedocs.io, moved them to a development branch so they no longer render on the website

And then @benjaminpope 's comments:

> This section of the paper needs a bit of work. It is maybe not 100% clear from reading it what `katsu` does and how it relates to other optics packages. I would suggest you have another half paragraph right at the top situating it in the context of other optics packages (eg you use dLux and prysm - cite and briefly say what these and any other relevant packages do). Then also explain, in one or two sentences, what Mueller matrices and Stokes vectors are - even just words to the effect that Stokes vectors are the appropriate representation for arbitrarily polarized light, and Mueller matrices are the objects that transform Stokes vectors when they propagate and interact with optical surfaces and media. It should be made clear, very briefly, that this is non-trivial. The existing text is too high level, and caters only to experts.

 Certainly, and thanks for the recommendations! dLux and prysm are not officially used now that I've descoped all of the "incomplete" tutorials to the dev_tutorials branch. Added a paragraph contextualizing the greater space of optical propagation software, including ray-tracing, physical optics (where I mention dLux and prysm), leading up to a description of Mueller calculus as a propagation technique.

> The intro to Mueller calculus section of the docs is good but still too high level - at the level of a couple of paragraphs and wiki links, nothing too onerous, I'd suggest that the docs explain what this package is for at a very basic level.

Added a paragraph with a discussion on the differences between Jones and Mueller calculus (included wikipedia links, thanks for that suggestion!). The landing page of the docs have been updated to contain a better descriptor of what the package does.

> The paper should probably explain a little more what PyPol is, and also situate this in the optics literature - the package makes use of both dLux and prysm in example notebooks, perhaps mention and describe this connection and these use cases?

Looks like I left a typo in the paper, it should be Pypolar*. Added a sentence to the paper on the common features between Katsu and Pypol, as well as routines that are unique to Pypolar. See previous comment on dLux and prysm not being officially used now that the incomplete tutorials are moved off of the main branch. I added them in a sentence discussion physical optics propagation packages.
