### 2025-02-18 #7375 @benjaminpope
Quick comments

- [ ] ? citation to ray-optics in the draft
- [ ] I would mention the _I, Q, U, V_ names for Stokes parameters as well as `s_i`
- [ ] I do like packages to have a quick access `katsu.__version__` attribute or similar so you can quickly verify installs

Everything else is fine and ready to sign off.
