### 2025-02-18 b5392b6
fixes in papers in response to reviewer comments
