+++ katsu/__init__.py
@@ -1,0 +1,2 @@
+__version__ = "0.2.0"
+__author__ = "Jaren N. Ashcraft"
