### 2025-02-26 #7375 @dfm
@Jashcraf — Thanks for all of your work so far! As part of the processing, can you:

1. Fix the invalid DOIs as suggested by the bot above ☝. A DOI shouldn't include the `https://doi.org/` prefix.
2. And update the `Dube2019` and `Dube:22` references so that the first author's name is consistent. You'll notice that in the paper these references are currently rendered as `B. Dube, 2019; B. D. Dube et al., 2022` in the text. We'd like this to instead just be `Dube, 2019; Dube et al., 2022`, but you can choose either "Brandon D. Dube" or "Brandon Dube" (unless these are different authors!).

Once you've done that, comment `@editorialbot generate pdf` in this thread and:
