+++ paper.bib
@@ -45,7 +45,7 @@ abstract = {The dual-rotating-retarder configuration is one of the most common f
   pages={A121},
   year={2023},
   publisher={EDP Sciences},
-  doi={	https://doi.org/10.1051/0004-6361/202245651}
+  doi={10.1051/0004-6361/202245651}
 }
 
 @software{poppy,
@@ -76,13 +76,13 @@ abstract = {The dual-rotating-retarder configuration is one of the most common f
 }
 
 @misc{gaudi2020habitable,
-      title={The Habitable Exoplanet Observatory (HabEx) Mission Concept Study Final Report}, 
+      title={The Habitable Exoplanet Observatory (HabEx) Mission Concept Study Final Report},
       author={B. Scott Gaudi and Sara Seager and Bertrand Mennesson and Alina Kiessling and Keith Warfield and Kerri Cahoy and John T. Clarke and Shawn Domagal-Goldman and Lee Feinberg and Olivier Guyon and Jeremy Kasdin and Dimitri Mawet and Peter Plavchan and Tyler Robinson and Leslie Rogers and Paul Scowen and Rachel Somerville and Karl Stapelfeldt and Christopher Stark and Daniel Stern and Margaret Turnbull and Rashied Amini and Gary Kuan and Stefan Martin and Rhonda Morgan and David Redding and H. Philip Stahl and Ryan Webb and Oscar Alvarez-Salazar and William L. Arnold and Manan Arya and Bala Balasubramanian and Mike Baysinger and Ray Bell and Chris Below and Jonathan Benson and Lindsey Blais and Jeff Booth and Robert Bourgeois and Case Bradford and Alden Brewer and Thomas Brooks and Eric Cady and Mary Caldwell and Rob Calvet and Steven Carr and Derek Chan and Velibor Cormarkovic and Keith Coste and Charlie Cox and Rolf Danner and Jacqueline Davis and Larry Dewell and Lisa Dorsett and Daniel Dunn and Matthew East and Michael Effinger and Ron Eng and Greg Freebury and Jay Garcia and Jonathan Gaskin and Suzan Greene and John Hennessy and Evan Hilgemann and Brad Hood and Wolfgang Holota and Scott Howe and Pei Huang and Tony Hull and Ron Hunt and Kevin Hurd and Sandra Johnson and Andrew Kissil and Brent Knight and Daniel Kolenz and Oliver Kraus and John Krist and Mary Li and Doug Lisman and Milan Mandic and John Mann and Luis Marchen and Colleen Marrese-Reading and Jonathan McCready and Jim McGown and Jessica Missun and Andrew Miyaguchi and Bradley Moore and Bijan Nemati and Shouleh Nikzad and Joel Nissen and Megan Novicki and Todd Perrine and Claudia Pineda and Otto Polanco and Dustin Putnam and Atif Qureshi and Michael Richards and A. J. Eldorado Riggs and Michael Rodgers and Mike Rud and Navtej Saini and Dan Scalisi and Dan Scharf and Kevin Schulz and Gene Serabyn and Norbert Sigrist and Glory Sikkia and Andrew Singleton and Stuart Shaklan and Scott Smith and Bart Southerd and Mark Stahl and John Steeves and Brian Sturges and Chris Sullivan and Hao Tang and Neil Taras and Jonathan Tesch and Melissa Therrell and Howard Tseng and Marty Valente and David Van Buren and Juan Villalvazo and Steve Warwick and David Webb and Thomas Westerhoff and Rush Wofford and Gordon Wu and Jahning Woo and Milana Wood and John Ziemer and Giada Arney and Jay Anderson and Jesús Maíz-Apellániz 
[…]
