### 2025-02-28 #7375 @Jashcraf
@dfm sent out to coauthors for a final check - will return when I hear their feedback, thanks again to @benjaminpope and @arendMoerman for the fantastic comments! Katsu is in much better shape now
