+++ paper.bib
@@ -259,7 +259,7 @@ title = {{Half-wave plate characterizations for the Keck NIRC2 polarimetry mode}
 organization = {International Society for Optics and Photonics},
 publisher = {SPIE},
 keywords = {NIRC2, polarimetric imaging, waveplate},
-year = {2023},
+year = {2024},
 doi = {10.1117/12.3019556}
 }
 
