+++ paper.md
@@ -47,7 +47,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-High-performance simulation of physical optics phenomena is instrumental to accurately design and understand optical systems. The propagation of light can be described in multiple regimes. The geometrical regime treats light as a ray, enabling the expeditious optimization of optical surfaces. Pythonic examples in the open source include Hayford et al.'s `ray-optics` [@ray-optics] and Myers et al.'s `batoid` [@batoid]. The physical regime treats light as a wave, enabling a precise understanding of the field distribution as light propagates through a given system. Many Python-based packages exist for wave optics propagation, including `POPPY` [@poppy], `HCIPy` [@hcipy], `prysm` [@Dube2019;@Dube:22], and `dLux` [@dLux]. However, all of the aforementioned packages treat light as a scalar field, and are unable to simulate the propagation of the vector nature of light. This property, called _polarization_, is critical for various instruments that interact with the vector nature of light.
+High-performance simulation of physical optics phenomena is instrumental in accurately designing and understanding optical systems. The propagation of light can be described in multiple regimes. The geometrical regime treats light as a ray, enabling the expeditious optimization of optical surfaces. Pythonic examples in the open source include Hayford et al.'s `ray-optics` [@ray-optics] and Myers et al.'s `batoid` [@batoid]. The physical regime treats light as a wave, enabling a precise understanding of the field distribution as light propagates through a given system. Many Python-based packages exist for wave optics propagation, including `POPPY` [@poppy], `HCIPy` [@hcipy], `prysm` [@Dube2019;@Dube:22], and `dLux` [@dLux]. However, all of the aforementioned packages treat light as a scalar field, and are unable to simulate the propagation of the vector nature of light. This property, called _polarization_, is critical for various instruments that interact with the vector nature of light.
 
 To describe the propagation of light's polarization state, we can use Mueller calculus. This approach represents optical systems as a 4 $\times$ 4 _Mueller Matrix_, ($\mathbf{M}$) that operate on a _Stokes Vector_  ($\mathbf{s}$) which represents the polarization of light, as shown in the Equation below,
 
@@ -67,7 +67,7 @@ $$
 
 Mueller calculus is particularly powerful because it is capable of describing light that is partially polarized, and the Stokes parameters are quantities that can be easily measured in the laboratory. 
 
-`Katsu` is an open-source Python package to address the need for polarimetric characterization of astronomical systems for the next generation of astronomical telescopes. It contains simple routines for the simulation of Mueller matrices and Stokes vectors to model how polarization is transformed by optical systems. This ability is not unique to `Katsu`; another package capable of simple Mueller calculus is the Pypolar [@pypol] package by Prahl, which also contains excellent visualization tools as well as support for ellipsometric data reduction. However, one area where `Katsu` is distinct from other packages capable of Mueller calculus simulation is its emphasis on broadcasted matrix calculations. All Mueller matrices available in `katsu.mueller` take a `shape` keword that append an arbitrary number of axes to the front of the initialized array, with the final two axes containing the Mueller matrix. This functionality is critical for accelerated computating on spatial data, which enables the direct measurement of polarization aberrations in the lab [@ashcraft2024]. `Katsu` also features a `polarimetry` module containing the data reduction routines for single-rotating retarder (SRR) Stokes polarimeters and dual-rotating retarder (DRR) Mueller polarimeters. In the pursuit of open-source instrumentation in the laboratory, `Katsu` supports an interface to the Newport Agilis series rotation mounts in the `motion` module to assist with performing polarimetry with rotating retarders. 
[…]
