### 2025-03-09 #7375 @dfm
Thanks @Jashcraf! If you're logged in to Zenodo, you should see and "Edit" button in the top right of the listing page:

Clicking on that will let you edit the metadata to match. Can you try that?

I'll get started with the final processing in the meantime.
