### 2025-03-09 #7375 @dfm
@Jashcraf — It also looks like you didn't fix the author formatting that I suggested above:

> And update the `Dube2019` and `Dube:22` references so that the first author's name is consistent. You'll notice that in the paper these references are currently rendered as `B. Dube, 2019; B. D. Dube et al., 2022` in the text. We'd like this to instead just be `Dube, 2019; Dube et al., 2022`, but you can choose either "Brandon D. Dube" or "Brandon Dube" (unless these are different authors!).

Can you fix that or otherwise comment on why my suggestion isn't correct? Thanks!
