+++ paper.bib
@@ -178,7 +178,7 @@ doi = {10.1117/12.2678001},
 URL = {https://doi.org/10.1117/12.2678001}
 }
 
-@article{Dube2019, doi = {10.21105/joss.01352}, url = {https://doi.org/10.21105/joss.01352}, year = {2019}, publisher = {The Open Journal}, volume = {4}, number = {37}, pages = {1352}, author = {Brandon Dube}, title = {prysm: A Python optics module}, journal = {Journal of Open Source Software} }
+@article{Dube2019, doi = {10.21105/joss.01352}, url = {https://doi.org/10.21105/joss.01352}, year = {2019}, publisher = {The Open Journal}, volume = {4}, number = {37}, pages = {1352}, author = {Brandon D. Dube}, title = {prysm: A Python optics module}, journal = {Journal of Open Source Software} }
 
 @article{Dube:22,
 author = {Brandon D. Dube and A. J. Riggs and Brian D. Kern and Eric J. Cady and John E. Krist and Hanying Zhou and Bijan Nemati and Byoung-Joon Seo and John Steeves and David Arndt and Milan Mandi\'{c} and Joel Shields and Dhemetrios Boussalis and Alfredo Valverde and Zahidul Rahman and Nanaz Fathpour},
