### 2025-03-12 #7375 @Jashcraf
@dfm author fixed in latest commit and metadata updated with authors / title synched up in Zenodo
https://zenodo.org/records/14963193
