### 2025-03-16 #7375 @warrickball
Hi @Jashcraf! I've started doing some preliminary checks ahead of the final publication steps and have a few comments.

1. I notice Ramya M. Anche is an author on the paper but _not_ listed as an author on the Zenodo repository. Unless I've missed it, could you explain why this is the case, or manually update the Zenodo author list?
2. Ashcraft et al. (2024) and Melby et al. (2024) appear to be missing conference titles, even though I'm pretty sure they're available through the SPIE pages linked through the DOIs. Could you complete the reference information here (and double check the other references while you're at it)?

Thanks! I may have more comments but most likely I'll open a pull request with editorial tweaks, after which I'll test the publication pipeline again.
