+++ docs/notebooks/incomplete/demo_spatial_calibration.py
@@ -1,0 +1,575 @@
+import matplotlib.pyplot as plt
+import matplotlib as mpl
+from scipy.optimize import minimize
+from tqdm import tqdm
+import numpy as tnp
+
+# Less common imports
+from prysm.coordinates import make_xy_grid, cart_to_polar
+from prysm.polynomials import noll_to_nm, sum_of_2d_modes
+from prysm.polynomials import zernike_nm_seq as zernike_nm_sequence
+from prysm.geometry import circle
+
+from katsu.mueller import (
+    linear_diattenuator,
+    linear_retarder,
+    linear_polarizer,
+    retardance_from_mueller
+)
+
+from katsu.polarimetry import drrp_data_reduction_matrix
+from katsu.katsu_math import np, set_backend_to_jax, broadcast_kron
+
+# Plotting stuff
+
+plt.style.use('bmh')
+okabe_colorblind8 = ['#E69F00','#56B4E9','#009E73',
+                     '#F0E442','#0072B2','#D55E00',
+                     '#CC79A7','#000000']
+plt.rcParams['axes.prop_cycle'] = mpl.cycler(color=okabe_colorblind8)
+
+
+def plot_square(x,n=4,vmin=None,vmax=None, common_cbar=True):
+    """A simple plotting macro for viewing intermediate results
+    """
+    k = 1
+    plt.figure(figsize=[10,10])
+    for i in range(n):
+        for j in range(n):
+            plt.subplot(n,n,k)
+            im = plt.imshow(x[..., i, j], vmin=vmin, vmax=vmax, cmap='RdBu_r')
+            if not common_cbar:
+                plt.colorbar()
+            k += 1
+
+    if common_cbar:
+        ax = plt.gca()
+        cbar = plt.colorbar(im, ax=ax)
+    
+    # plt.show()
+
+
+def jax_sum_of_2d_modes(modes, weights):
+    """a clone of prysm.polynomials sum_of_2d_modes that works when using katsu's Jax backend
+
+    Parameters
+    ----------
+    modes : list of ndarrays
+        list of polynomials constituting the desired basis
+    weights : list of floats
+        coefficients that describe the presence of each mode in the modes list
+
+    Returns
+    -------
+    ndarray
+        2D ndarray containing the sum of weighted modes
+    """
+    modes = np.asarray(modes)
+    weights = np.asarray(weights).astype(modes.dtype)
+
+    # dot product of the 0th dim of modes and weights => weighted sum
+    return np.tensordot(modes, weights, axes=(0, 0))
+
+def sum_of_2d_modes_wrapper(modes, weights):
+    """ Wrapper that lets us ignore which source module we want to use
+    """
+    if np._srcmodule == tnp:
+        return sum_of_2d_modes(modes, weights)
+    else:
+        return jax_sum_of_2d_modes(modes, weights)
+
+
+def construct_zern_basis(r, t, NMODES):
+    """Builds a prysm zernike basis
+
+    Parameters
+    ----------
+    r : ndarray
+        radial coordinate
+    t : ndarray
+        azimuthal coordinate
+    NMODES: int
+        Maximum zernike noll index to simulate 
+
+    Returns
+    -------
+    list
+        list whose elements contain Zernike modes
+    """
+
[…]
