+++ paper.bib
@@ -257,7 +257,7 @@ year = {2025}
 author = {William Melby and Manxuan Zhang and Maxwell Millar-Blanchaer and Anthony D. Manni and Douglas S. Hobbs and Jaren Ashcraft},
 title = {{Half-wave plate characterizations for the Keck NIRC2 polarimetry mode}},
 organization = {International Society for Optics and Photonics},
-publisher = {SPIE},
+publisher = {SPIE Astronomical Telescopes + Instrumentation 2024},
 keywords = {NIRC2, polarimetric imaging, waveplate},
 year = {2024},
 doi = {10.1117/12.3019556}
