### 2025-03-21 #7375 @Jashcraf
Hi @warrickball , thanks for the catches!

Added Ramya M. Anche to author list in Zenodo repo: https://zenodo.org/records/14963193

Updated publisher in Melby et al. (2024) to be SPIE Astronomical Telescopes + Instrumentation 2024, which was the name of the conference. The Ashcraft et al. (2024) is actually an Arxiv posting of the SPIE Proceedings, do I still need to include the conference information for this bib entry?
