### 2025-03-23 57ece5b
Initial editorial tweaks for JOSS paper
