+++ paper.md
@@ -5,8 +5,8 @@ tags:
   - astronomy
   - polarimetry
   - laboratory
-  - mueller
-  - stokes
+  - Mueller
+  - Stokes
 authors:
   - name: Jaren N. Ashcraft
     orcid: 0000-0001-5082-7442
@@ -32,7 +32,7 @@ authors:
   - name: Maxwell A. Millar-Blanchaer
     orcid: 0000-0001-6205-9233
     affiliation: 3
-  
+
 affiliations:
  - name: Wyant College of Optical Sciences, University of Arizona, USA
    index: 1
@@ -47,9 +47,9 @@ bibliography: paper.bib
 ---
 
 # Summary
-High-performance simulation of physical optics phenomena is instrumental in accurately designing and understanding optical systems. The propagation of light can be described in multiple regimes. The geometrical regime treats light as a ray, enabling the expeditious optimization of optical surfaces. Pythonic examples in the open source include Hayford et al.'s `ray-optics` [@ray-optics] and Myers et al.'s `batoid` [@batoid]. The physical regime treats light as a wave, enabling a precise understanding of the field distribution as light propagates through a given system. Many Python-based packages exist for wave optics propagation, including `POPPY` [@poppy], `HCIPy` [@hcipy], `prysm` [@Dube2019;@Dube:22], and `dLux` [@dLux]. However, all of the aforementioned packages treat light as a scalar field, and are unable to simulate the propagation of the vector nature of light. This property, called _polarization_, is critical for various instruments that interact with the vector nature of light.
+High-performance simulation of physical optics phenomena is instrumental in accurately designing and understanding optical systems. The propagation of light can be described in multiple regimes. The geometrical regime treats light as a ray, enabling the expeditious optimization of optical surfaces. Pythonic examples in open source include `ray-optics` [@ray-optics] and `batoid` [@batoid]. The physical regime treats light as a wave, enabling a precise understanding of the field distribution as light propagates through a given system. Many Python-based packages exist for wave optics propagation, including `POPPY` [@poppy], `HCIPy` [@hcipy], `prysm` [@Dube2019;@Dube:22], and [$\partial$Lux](https://github.com/louisdesdoigts/dLux) [@dLux]. However, all of the aforementioned packages treat light as a scalar field, and are unable to simulate the propagation of the vector nature of light. This property, called _polarization_, is critical for various instruments that interact with the vector nature of light.
 
-To describe the propagation of light's polarization state, we can use Mueller calculus. This approach represents optical systems as a 4 $\times$ 4 _Mueller Matrix_, ($\mathbf{M}$) that operate on a _Stokes Vector_  ($\mathbf{s}$) which represents the polarization of light, as shown in the Equation below,
+To describe the propagation of light's polarization state, we can use Mueller calculus. This approach represents optical systems as a 4 $\times$ 4 _Mueller matrix_ $\mathbf{M}$ that operates on a _Stokes vector_ $\mathbf{s}$, which represents the polarization of light, as shown in the equation,
 
 $$\mathbf{s}' = \mathbf{M} \mathbf{s}.$$
 
@@ -65,20 +65,20 @@ $$
 \mathbf{s} = [I, Q, U, V].
 $$
 
-Mueller calculus is particularly powerful because it is capable of describing light that is partially polarized, and the Stokes parameters are quantities that can be easily measured in the laboratory. 
+Mueller calculus is particularly powerful because it is capable of describing light that is partially polarized, and the Stokes parameters are quantities that can be easily measured in the laboratory.
 
[…]
