### 2025-03-23 edf4827
tweak casing of PyAstroPol in text
