+++ paper.md
@@ -73,7 +73,7 @@ To interpret the measurements made in the lab, Katsu features the polar decompos
 
 $$\mathbf{M} = \mathbf{M}_{\Delta}\mathbf{M_{R}}\mathbf{M_{D}}. $$
 
-This function is critical for separating depolarization from the Mueller matrix, which enables the integration of polarization aberration in the laboratory into simulated data generated by a polarization ray tracer, e.g. `Poke` [@Ashcraft_2023], `Pyastropol` [@Pruthvi.2020]. Katsu also adopts the flexible interchangeable backend system of `prysm` [@Dube2019;@Dube:22] for hot-swappable NumPy-like backends [@harris2020array] including CuPy for GPU-accelerated computing [@cupy_learningsys2017] and JAX for automatic differentiation [@jax2018github].
+This function is critical for separating depolarization from the Mueller matrix, which enables the integration of polarization aberration in the laboratory into simulated data generated by a polarization ray tracer, e.g. `Poke` [@Ashcraft_2023], `PyAstroPol` [@Pruthvi.2020]. Katsu also adopts the flexible interchangeable backend system of `prysm` [@Dube2019;@Dube:22] for hot-swappable NumPy-like backends [@harris2020array] including CuPy for GPU-accelerated computing [@cupy_learningsys2017] and JAX for automatic differentiation [@jax2018github].
 
 # Statement of need
 Polarimetric characterization in the laboratory is critical to the performance of astronomical instrumentation. The next generation of astronomical observatories has identified that wavefront errors induced by polarization, called _polarization aberrations_, could be a limiting factor in direct exoplanet imaging. Ample modeling has been done at the observatory level to understand the nominal polarization aberrations that arise in the telescope [@anche2023polarization;@gaudi2020habitable;@Will_polarization_luvoir] but less work has been done to characterize instrumentation in the laboratory.
