### 2025-03-23 #7375 @warrickball
I've opened a [pull request](https://github.com/Jashcraf/katsu/pull/38) with some editorial tweaks. If you're happy with those, the I'll give the publication pipeline another test run once the changes are merged and check again if anything else needs tweaking.
