### 2025-03-24 d355c58
Drop hyperlink ∂Lux to fix JOSS publication pipeline
