### 2025-03-25 #7375 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
