+++ paper.md
@@ -67,7 +67,7 @@ $$
 
 Mueller calculus is particularly powerful because it is capable of describing light that is partially polarized, and the Stokes parameters are quantities that can be easily measured in the laboratory.
 
-Katsu is an open-source Python package to address the need for polarimetric characterization of astronomical systems for the next generation of astronomical telescopes. It contains simple routines for the simulation of Mueller matrices and Stokes vectors to model how polarization is transformed by optical systems. This ability is not unique to Katsu. Another package capable of simple Mueller calculus is [Pypolar](https://github.com/scottprahl/pypolar) [@pypol], which also contains excellent visualization tools as well as support for ellipsometric data reduction. However, one area where Katsu is distinct from other packages capable of Mueller calculus simulation is its emphasis on broadcasted matrix calculations. All Mueller matrices available in `katsu.mueller` take a `shape` keyword that appends an arbitrary number of axes to the front of the initialized array, with the final two axes containing the Mueller matrix. This functionality is critical for accelerated computing on spatial data, which enables the direct measurement of polarization aberrations in the lab [@ashcraft2024]. Katsu also features a `polarimetry` module containing the data reduction routines for single-rotating retarder (SRR) Stokes polarimeters and dual-rotating retarder (DRR) Mueller polarimeters. In the pursuit of open-source instrumentation in the laboratory, Katsu supports an interface to the Newport Agilis series rotation mounts in the `motion` module to assist with performing polarimetry with rotating retarders.
+Katsu is an open-source Python package to address the need for polarimetric characterization of astronomical systems for the next generation of astronomical telescopes. It contains simple routines for the simulation of Mueller matrices and Stokes vectors to model how polarization is transformed by optical systems. This ability is not unique to Katsu. Another package capable of simple Mueller calculus is [pypolar](https://github.com/scottprahl/pypolar) [@pypol], which also contains excellent visualization tools as well as support for ellipsometric data reduction. However, one area where Katsu is distinct from other packages capable of Mueller calculus simulation is its emphasis on broadcasted matrix calculations. All Mueller matrices available in `katsu.mueller` take a `shape` keyword that appends an arbitrary number of axes to the front of the initialized array, with the final two axes containing the Mueller matrix. This functionality is critical for accelerated computing on spatial data, which enables the direct measurement of polarization aberrations in the lab [@ashcraft2024]. Katsu also features a `polarimetry` module containing the data reduction routines for single-rotating retarder (SRR) Stokes polarimeters and dual-rotating retarder (DRR) Mueller polarimeters. In the pursuit of open-source instrumentation in the laboratory, Katsu supports an interface to the Newport Agilis series rotation mounts in the `motion` module to assist with performing polarimetry with rotating retarders.
 
 To interpret the measurements made in the lab, Katsu features the polar decomposition of Mueller matrices published by Lu and Chipman [@Lu:96]. This decomposes a Mueller matrix $\mathbf{M}$ into its constituent depolarizer $\mathbf{M}_{\Delta}$, diattenuator $\mathbf{M_{D}}$, and retarder $\mathbf{M_{R}}$, as shown in the following equation,
 
