### 2025-03-28 #7375 @warrickball
Hi @Jashcraf, I noticed a few more tiny tweaks (sorry for not catching them before) and opened [another PR](https://github.com/Jashcraf/katsu/pull/40). If we merge that I think we should be ready to publish.
