### 2025-03-29 #7375 @warrickball
Congratulations @Jashcraf, your paper has been published in JOSS! 🎉
